+++ service/adapters/badger/ban_list_repository_test.go
@@ -4,11 +4,11 @@ import (
 	"bytes"
 	"testing"
 
-	"github.com/planetary-social/scuttlego/di"
 	"github.com/planetary-social/scuttlego/internal"
 	"github.com/planetary-social/scuttlego/internal/fixtures"
 	badgeradapters "github.com/planetary-social/scuttlego/service/adapters/badger"
 	"github.com/planetary-social/scuttlego/service/app/commands"
+	"github.com/planetary-social/scuttlego/service/di"
 	"github.com/planetary-social/scuttlego/service/domain/bans"
 	"github.com/stretchr/testify/require"
 )