+++ service/adapters/badger/message_repository_test.go
@@ -3,9 +3,9 @@ package badger_test
 import (
 	"testing"
 
-	"github.com/planetary-social/scuttlego/di"
 	"github.com/planetary-social/scuttlego/internal/fixtures"
 	"github.com/planetary-social/scuttlego/service/adapters/badger"
+	"github.com/planetary-social/scuttlego/service/di"
 	"github.com/stretchr/testify/require"
 )