+++ service/adapters/badger/receive_log_repository_test.go
@@ -4,11 +4,11 @@ import (
 	"testing"
 
 	"github.com/boreq/errors"
-	"github.com/planetary-social/scuttlego/di"
 	"github.com/planetary-social/scuttlego/internal/fixtures"
 	"github.com/planetary-social/scuttlego/service/adapters/badger"
 	"github.com/planetary-social/scuttlego/service/app/common"
 	"github.com/planetary-social/scuttlego/service/app/queries"
+	"github.com/planetary-social/scuttlego/service/di"
 	"github.com/stretchr/testify/require"
 )